+++ src/shadow-dom.js
@@ -1,11 +1,13 @@
-import {inject} from 'aurelia-dependency-injection';
 import {DOM} from 'aurelia-pal';
 import {_isAllWhitespace} from './util';
 
 let noNodes = Object.freeze([]);
 
-@inject(DOM.Element)
 export class SlotCustomAttribute {
+  static inject() {
+    return [DOM.Element];
+  }
+
   constructor(element) {
     this.element = element;
     this.element.auSlotAttribute = this;